+++ js/functions.js
@@ -5,8 +5,6 @@ const checkValidLength = function (string, maxLength) {
   return string.length <= maxLength;
 };
 
-checkValidLength('проверяем эту строку', 10);
-checkValidLength('проверяем эту строку', 20);
 checkValidLength('проверяем эту строку', 22);
 
 
@@ -23,9 +21,6 @@ const isPalindrome = function (string) {
   return normalizedString === newString;
 };
 
-isPalindrome('топот');
-isPalindrome('ДовОд ');
-isPalindrome('Кекс');
 isPalindrome('Лёша на полке клопа нашёл ');
 
 //Функция проверяет является ли строка палиндромом: true / false. Вариант 2
@@ -41,9 +36,6 @@ const isPalindrome2 = function (string) {
   return true;
 };
 
-isPalindrome2('топот');
-isPalindrome2('ДовОд ');
-isPalindrome2('Кекс');
 isPalindrome2('Лёша на полке клопа нашёл ');
 
 //Функция проверяет является ли строка палиндромом: true / false. Вариант 3
@@ -87,13 +79,6 @@ const extractDigits = function (parameter) {
 };
 
 extractDigits('2023 год');
-extractDigits('ECMAScript 2022');
-extractDigits('1 кефир, 0.5 батона');
-extractDigits('агент 007');
-extractDigits('а я томат');
-extractDigits(2023);
-extractDigits(-1);
-extractDigits(1.5);
 
 
 //Функция извлекает цифры из строки и возвращает в виде целого положительного числа.
@@ -120,10 +105,3 @@ const extractDigits2 = function (parameter) {
 };
 
 extractDigits2('2023 год');
-extractDigits2('ECMAScript 2022');
-extractDigits2('1 кефир, 0.5 батона');
-extractDigits2('агент 007');
-extractDigits2('а я томат');
-extractDigits2(2023);
-extractDigits2(-1);
-extractDigits2(1.5);
