+++ package.json
@@ -16,6 +16,7 @@
     "eslint-config-htmlacademy": "9.1.1"
   },
   "license": "MIT",
+
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
@@ -23,5 +24,6 @@
   "engines": {
     "node": "18",
     "npm": "9"
-  }
+  },
+  "type": "module"
 }
